add debounce and create helpers
