+++ components/container/ObserverSections.js
@@ -1,6 +1,7 @@
 import React, { useState, useEffect } from 'react';
 import PropTypes from 'prop-types';
 import { ObserverSection } from 'react-components';
+import { indexOfMax } from 'react-components';
 
 const ObserverSections = ({ list }) => {
     const [intersectionRatios, setIntersectionRatios] = useState(
@@ -10,8 +11,6 @@ const ObserverSections = ({ list }) => {
     );
 
     useEffect(() => {
-        // find id corresponding to biggest intersection ratio
-        const indexOfMax = (arr) => arr.findIndex((elm) => elm === Math.max(...arr));
         const idOfMax = list[indexOfMax(intersectionRatios)].id;
 
         // replace URL
