+++ index.js
@@ -18,6 +18,8 @@ export { default as Bordered } from './components/container/Bordered';
 export { default as Block } from './components/container/Block';
 export { default as Row } from './components/container/Row';
 export { default as Information } from './components/container/Information';
+export { default as ObserverSection } from './components/container/ObserverSection';
+export { default as ObserverSections } from './components/container/ObserverSections';
 export { default as Tooltip } from './components/tooltip/Tooltip';
 export { default as PrimaryButton } from './components/button/PrimaryButton';
 export { default as WarningButton } from './components/button/WarningButton';
@@ -77,8 +79,6 @@ export { default as Label } from './components/label/Label';
 export { default as NavMenu } from './components/sidebar/NavMenu';
 export { default as NavItem } from './components/sidebar/NavItem';
 export { default as Sidebar } from './components/sidebar/Sidebar';
-export { default as ObserverSection } from './components/sidebar/ObserverSection';
-export { default as ObserverSections } from './components/sidebar/ObserverSections';
 export { default as LoginContainer } from './containers/login/LoginContainer';
 export { default as TOTPForm } from './containers/login/TOTPForm';
 export { default as useLogin } from './containers/login/useLogin';
@@ -220,6 +220,7 @@ export { default as useApiResult } from './hooks/useApiResult';
 export { default as useAuthenticationStore } from './hooks/useAuthenticationStore';
 export { default as useAsync } from './hooks/useAsync';
 export { nestChildren } from './helpers/component';
+export { indexOfMax, buildThresholds } from './helpers/intersectionObserver';
 
 export { default as useCache } from './containers/cache/useCache';
 export { default as useCachedResult } from './containers/cache/useCachedResult';
